pyproject.toml
[tool.poetry]
name = "semaphore-api"
version = "0.1.0"
description = "Ansible modules for Ansible Semaphore API"
authors = ["Vítězslav Dvořák <info@vitexsoftware.cz>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.11"
ansible = "^7.1.0"
ansible-semaphore-client = "^0.0.1"

[tool.poetry.dev-dependencies]
ansible-lint = "^6.9.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
